Refactor and organize code
